requirements.txt
sentence-transformers
torch
numpy
pandas
scikit-learn
faiss-cpu
tqdm
click
flask
werkzeug